Create 1 Task: Scrabble.c
This task of week 2 in CS50 course, was to create a program that resembles the table top game scrabble. Two players will compete against each other typing a word each, then the program will calculate the scores of each character from an array and pronounce the winner in the end.