Create 1 Task: Caesar.c
This challenge was to be able to encrypt a piece of text given by the user, using Caesar's encryption algorithm. It requires the user to first give a whole number input, as well as a piece of text after which it will encrypt it based on the key value given earlier. For future, it could also be worth quickly implement a range for max and min value, so that the program won't crash as the bits is limited.